model/requirements.txt
llama-cpp-python
fastapi
uvicorn
pydantic